text-bison-001-tagger/requirements.txt
google-generativeai
time
sys
dotenv
os